desktop: Refactor gui controller into its own file
